+++ x-pack/plugins/apm/server/routes/diagnostics/route.ts
@@ -7,13 +7,11 @@
 
 import { IndicesDataStream } from '@elastic/elasticsearch/lib/api/types';
 import { createApmServerRoute } from '../apm_routes/create_apm_server_route';
-import {
-  getMatchingIndexTemplates,
-  getApmIndexPatterns,
-} from './index_templates/get_matching_index_templates';
 import { getDefaultApmIndexTemplateStates } from './index_templates/get_default_apm_index_templates_states';
 import { getIndicesWithStatuses } from './indices/get_indices';
 import { getApmIndices } from '../settings/apm_indices/get_apm_indices';
+import { getIndexTemplatesByIndexPattern } from './index_templates/get_index_pattern_settings';
+import { getDataStreams } from './data_streams/get_data_streams';
 
 export interface IndiciesItem {
   index: string;
@@ -73,12 +71,9 @@ const indexPatternSettingsRoute = createApmServerRoute({
       config: resources.config,
     });
 
-    const defaultApmIndexTemplateStates =
-      await getDefaultApmIndexTemplateStates({ esClient });
-    const indexTemplatesByIndexPattern = await getMatchingIndexTemplates({
-      apmIndices,
+    const indexTemplatesByIndexPattern = await getIndexTemplatesByIndexPattern({
       esClient,
-      defaultApmIndexTemplateStates,
+      apmIndices,
     });
 
     return { indexTemplatesByIndexPattern };
@@ -115,34 +110,13 @@ const dataStreamRoute = createApmServerRoute({
     nonDataStreamIndices: string[];
   }> => {
     const coreContext = await resources.context.core;
-    const { asCurrentUser } = coreContext.elasticsearch.client;
-    const apmIndexPatterns = getApmIndexPatterns(
-      await getApmIndices({
-        savedObjectsClient: coreContext.savedObjects.client,
-        config: resources.config,
-      })
-    );
-
-    // fetch APM data streams
-    const { data_streams: dataStreams } =
-      await asCurrentUser.indices.getDataStream({
-        name: apmIndexPatterns,
-        filter_path: ['data_streams.name', 'data_streams.template'],
-      });
-
-    // fetch non-data stream indices
-    const nonDataStreamIndicesResponse = await asCurrentUser.indices.get({
-      index: apmIndexPatterns,
-      filter_path: ['*.data_stream', '*.settings.index.uuid'],
+    const { asCurrentUser: esClient } = coreContext.elasticsearch.client;
+    const apmIndices = await getApmIndices({
+      savedObjectsClient: coreContext.savedObjects.client,
+      config: resources.config,
     });
 
-    const nonDataStreamIndices = Object.entries(nonDataStreamIndicesResponse)
-      .filter(
-        ([indexName, { data_stream: dataStream }]): boolean => !dataStream
-      )
-      .map(([indexName]): string => indexName);
-
-    return { dataStreams, nonDataStreamIndices };
+    return getDataStreams({ apmIndices, esClient });
   },
 });
